move credentials to extra class
